+++ src/components/FooterBar.js
@@ -1,0 +1,16 @@
+import '../App.css';
+import '../styles.css';
+
+function FooterBar() {
+    return (
+      <nav className="menubar">
+        <div>
+          <button>Contact</button>
+          <button>Careers</button>
+          <button>Help</button>
+        </div>
+      </nav> 
+    )
+  }
+
+export {FooterBar};
